+++ HISTORY.rst
@@ -3,6 +3,14 @@
 History
 -------
 
+2.11.0
++++++++++++++++++++
+
+* Updated the validation for the Report Transactions API to make the
+  ``ip_address`` parameter optional. Now the ``tag`` and at least one of the
+  following parameters must be supplied: ``ip_address``, ``maxmind_id``,
+  ``minfraud_id``, ``transaction_id``.
+
 2.10.0 (2024-04-16)
 +++++++++++++++++++